+++ packages/socket/test/socket_spec.js
@@ -114,6 +114,41 @@ describe('Socket', function () {
       }
     })
 
+    it('correctly encodes and decodes circular data in array', (done) => {
+      const encoder = new parser.Encoder()
+
+      const circularObj = {
+        foo: {},
+      }
+
+      circularObj.foo.circularArray = [circularObj, circularObj]
+
+      const obj = {
+        type: PacketType.EVENT,
+        data: ['a', circularObj],
+        id: 23,
+        nsp: '/cool',
+      }
+
+      const originalData = obj.data
+
+      const encodedPackets = encoder.encode(obj)
+
+      const decoder = new parser.Decoder()
+
+      decoder.on('decoded', (packet) => {
+        obj.data = originalData
+        expect(packet.data[1] === packet.data[1].foo.circularArray[0]).to.be.true
+        expect(packet.data[1] === packet.data[1].foo.circularArray[1]).to.be.true
+        expect(packet).to.eql(obj)
+        done()
+      })
+
+      for (let i = 0; i < encodedPackets.length; i++) {
+        decoder.add(encodedPackets[i])
+      }
+    })
+
     it('correctly encodes and decodes circular data containing binary', (done) => {
       const encoder = new parser.Encoder()